rating fix (parse to float instead of int)

+++ src/application/endpoints/guide.endpoint.ts
@@ -84,7 +84,7 @@ export class GuideEndpoint {
     @Validate(query('rating','a rating has to be defined').isNumeric())
     async getByRating(@Query('rating') rating: string, req: Request, res: Response){
         try{
-            const guides: IGuide[] = await this.unitOfWork.guides.getByRating(Number.parseInt(rating));
+            const guides: IGuide[] = await this.unitOfWork.guides.getByRating(Number.parseFloat(rating));
             return Ok(this.convertToDto(guides));
         } catch(err){
             return BadRequest({msg: err.message});
